module6-task1 *Рефакторинг* Убрал лишний код из PageController, отвечающий за отрисовку компонента карточки фильма и перенес всю логику по отрисовке карточки, попап окна с фильмом и его закрытие в FilmController
